Reorganized dashboard page tab components
