+++ src/views/DashboardView/Tabs/index.ts
@@ -1,0 +1,5 @@
+import { DetailTab } from './DetailTab';
+import { HourlyForecastTab } from './HourlyForecastTab';
+import { DailyForecastTab } from './DailyForecastTab';
+
+export { DetailTab, HourlyForecastTab, DailyForecastTab };
